% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajusteBinomialNegativa.R
\name{ajusteBinomialNegativa}
\alias{ajusteBinomialNegativa}
\title{Ajusta modelos com resposta BinomialNegativa(\eqn{\mu},\eqn{\phi})}
\usage{
ajusteBinomialNegativa(
  y,
  X,
  init,
  phi_init,
  eps = 1e-06,
  maxiter = 50,
  gamma = 0.95,
  f,
  dados
)
}
\arguments{
\item{y}{`vector` com a contagem de eventos (números
inteiros);}

\item{X}{`matrix` com as variáveis preditoras;}

\item{init}{`matrix` com \eqn{p} linhas e uma coluna com os
chutes iniciais dos valores dos \eqn{p} parâmetros a serem
 estimados;}

\item{phi_init}{`numeric` com o chute inicial do parâmetro
de dispersão \eqn{\phi} a ser estimado;}

\item{eps}{Valor de \eqn{\varepsilon}, a distância máxima
entre duas iterações consecutivas do método de Newton para
se dizer que houve convergência do método;}

\item{maxiter}{Número máximo de iterações a serem realizadas
no método de Newton para obtenção das estimativas dos
parâmetros;}

\item{gamma}{Variável do tipo 'numeric' entre 0 e 1, indicando a confiança desejada no intervalo a ser construído para os parâmetros.}

\item{link}{Função de ligação a ser utilizada no modelo. As
funções de ligação implementadas (e a forma de utilizá-las)
até o momento são:
\itemize{
\item{"logit": }{Ligação logística, dada por
 \eqn{g(\mu) = \log\left(\frac{\mu}{1-\mu} \right);}
\item{"probit": }{Ligação probito, a inversa da f.d.a. da
 normal padrão, dada por \eqn{g(\mu) = \phi^{-1}(\mu)},
 onde \eqn{\phi(.)} denota o valor da f.d.a. de uma normal
 padrão avaliada em (.);}
 \item {"cloglog": } Ligação log-log complementar, dada por
 \eqn{g(1-\mu) = \log(-\log(1-\mu))};}
}}
}
\value{
Retorna uma lista com as informações do ajuste. A
descrição de cada item está apresentada abaixo:
\itemize{
\item out: `matrix` com as estimativas dos parâmetros geradas
 a cada iteração do método de Newton;
\item est: `matrix` com as estimativas (finais) dos parâmetros
  do modelo;
\item iter: Número de iterações necessárias até a convergência;
\item se: `vector` com os erros-padrão dos estimadores dos
parâmetros do modelo;
\item preditor: `matrix` com o valor de \eqn{\eta} (preditor
 linear) de cada observação;
\item v.ajust: `matrix` com os valores das médias ajustadas
para cada observação;
\item z.value: `matrix` com os valores das estatísticas-Z
do teste individual para cada \eqn{\beta_i = 0};
\item p.value: `matrix` com os p-valores dos testes individuais para cada \eqn{\beta_i = 0  \times  \beta_i \neq 0};
\item LI: `matrix` com os limites inferiores dos intervalos de confiança de \eqn{\gamma}% para os parâmetros;
\item LS: `matrix` com os limites superiores dos intervalos de confiança de \eqn{\gamma}% para os parâmetros;
\item AIC: AIC do modelo ajustado;
\item BIC: BIC do modelo ajustado;
\item desvio: Desvio do modelo ajustado;
\item gl.res: Número de graus de liberdade dos resíduos;
\item comp.desvio: Resíduos componentes do desvio do modelo
 ajustado;
\item residuos.pearson: Resíduos de Pearson do desvio do modelo
 ajustado;
\item comp.desvio.st: Resíduos componentes do desvio padronizados
 do modelo ajustado;
\item residuos.pearson.st: Resíduos de Pearson padronizados
 do modelo ajustado;
 \item Hhat: Matriz hessiana ajustada.
}
}
\description{
Esta função auxiliar ajusta modelos de regressão com resposta
Binomial Negativa. É possível (até o momento) utilizar
 apenas a função de ligação `log`.
}
