% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlg.R
\name{mlg}
\alias{mlg}
\title{Ajusta modelos lineares generalizados (MLG)}
\usage{
mlg(
  f,
  dados = NULL,
  init,
  phi_init,
  link = "logit",
  distr,
  eps = 1e-06,
  maxiter = 100,
  show.result = TRUE,
  gamma = 0.95
)
}
\arguments{
\item{f}{Fórmula do ajuste a ser realizado;}

\item{dados}{Conjunto de dados do tipo `data.frame` com as colunas especificadas na fórmula (parâmetro `f`);}

\item{init}{`matrix` com uma coluna e `p` linhas, onde `p` é o
número de parâmetros incluídos no modelo. Esses valores serão
 utilizados como chutes iniciais para o ajuste via método de
 Newton-Raphson;}

\item{phi_init}{objeto do tipo `numeric` com o valor inicial
para o parâmetro \eqn{\phi} quando se está estimando um modelo
binomial negativo;}

\item{link}{Função de ligação a ser utilizada no modelo. As funções de ligação implementadas até o momento são:
\itemize{
 \item{Binomial: }{"logit" (ligação logística), "probit" (ligação
  probito, a inversa da f.d.a. da normal padrão) e "cloglog" (
  ligação log-log complementar);}
 \item{Binomial negativa: }{"log"}
 \item{Poisson: }{"log"}
}}

\item{distr}{Distribuição assumida para a variável resposta. Até
o momento, estão implementadas as distribuições "Binomial",
"Binomial negativa" e "Poisson";}

\item{eps}{Valor de \eqn{\varepsilon}, a distância máxima entre
duas iterações consecutivas do método de Newton para se dizer
que houve convergência do método;}

\item{maxiter}{Número máximo de iterações a serem realizadas no
método de Newton para obtenção das estimativas dos parâmetros;}

\item{show.result}{Variável do tipo `logical` indicando se é
de interesse imprimir os valores dos parâmetros e outras
informações sumárias;}

\item{gamma}{Variável do tipo `numeric` entre 0 e 1, indicando
a confiança desejada no intervalo a ser construído para os
parâmetros.}
}
\value{
Retorna uma lista com as informações do ajuste. A
descrição de cada item está apresentada abaixo:
\itemize{
\item out: `matrix` com as estimativas dos parâmetros geradas
 a cada iteração do método de Newton;
\item est: `matrix` com as estimativas (finais) dos parâmetros
  do modelo;
\item iter: Número de iterações necessárias até a convergência;
\item se: `vector` com os erros-padrão dos estimadores dos
parâmetros do modelo;
\item preditor: `matrix` com o valor de \eqn{\eta} (preditor
 linear) de cada observação;
\item mv.ajust: `matrix` com os valores das médias ajustadas
para cada observação;
\item z.values: `matrix` com os valores das estatísticas-Z
do teste individual para cada \eqn{\beta_i = 0};
\item p.value: `matrix` com os p-valores dos testes individuais para cada \eqn{\beta_i = 0  \times  \beta_i \neq 0};
\item LI: `matrix` com os limites inferiores dos intervalos de confiança de \eqn{\gamma}% para os parâmetros;
\item LS: `matrix` com os limites superiores dos intervalos de confiança de \eqn{\gamma}% para os parâmetros;
\item AIC: AIC do modelo ajustado;
\item BIC: BIC do modelo ajustado;
\item desvio: Desvio do modelo ajustado;
\item gl.res: Número de graus de liberdade dos resíduos;
\item comp.desvio: Resíduos componentes do desvio do modelo
 ajustado;
\item residuos.pearson: Resíduos de Pearson do desvio do modelo
 ajustado;
\item comp.desvio.st: Resíduos componentes do desvio padronizados
 do modelo ajustado;
\item residuos.pearson.st: Resíduos de Pearson padronizados
 do modelo ajustado;
 \item Hhat: Matriz hessiana ajustada.
}
}
\description{
A função realiza o ajuste de vários modelos lineares generalizados
manualmente, retornando um intervalo de confiança para os
parâmetros estimados, além dos objetos retornados pela função
`glm`.
}
